+++ src/PdfPinata/Drawing/XBitmapImage.cs
@@ -34,8 +34,6 @@ namespace PdfPinata.Drawing;
 /// </summary>
 public sealed class XBitmapImage : XBitmapSource
 {
-    // TODO: Move code from XImage to this class.
-
     /// <summary>
     /// Initializes a new instance of the <see cref="XBitmapImage"/> class.
     /// </summary>
@@ -60,6 +58,7 @@ public sealed class XBitmapImage : XBitmapSource
     /// <summary>
     /// Creates a default 24 bit ARGB bitmap with the specified pixel size.
     /// </summary>
+    [System.Obsolete("The bitmap has a size and no pixels, and nothing can draw onto it. It will be removed.")]
     public static XBitmapSource CreateBitmap(int width, int height)
     {
         // Create a default 24 bit ARGB bitmap.